Move files and deleted blank fragment
