--- app/src/main/res/layout/fragment_blank.xml
@@ -1,15 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="org.cpen321.discovr.BlankFragment">
-
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:textColor="@color/primaryTextColor"
-        android:text="@string/hello_blank_fragment"
-        android:background="@color/colorPrimaryDark"/>
-
-</FrameLayout>
